pyproject.toml
[tool.poetry]
name = "HeroCrafter"
version = "0.1.0"
description = ""
authors = ["Andrey Lisin <andrey.lisin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
openai = "^1.30.5"
django-bootstrap4 = "^24.3"
psycopg2-binary = "^2.9.9"
uwsgi = "^2.0.26"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
